test: avoid server extras in memory filter unit tests
