+++ tests/unit/test_list_memory_filters.py
@@ -1,12 +1,8 @@
 from unittest.mock import MagicMock
 
-import pytest
-from starlette.requests import Request
-
 from powermem.core.memory import Memory
 from powermem.storage.adapter import StorageAdapter
 from powermem.storage.sqlite.sqlite_vector_store import SQLiteVectorStore
-from server.api.v1.memories import list_memories
 from server.services.memory_service import MemoryService
 
 
@@ -103,63 +99,4 @@ def test_memory_service_list_and_count_pass_filters():
         user_id="u01",
         agent_id="a01",
         filters={"scope": "group"},
-    )
-
-
-@pytest.mark.asyncio
-async def test_list_memories_api_converts_scope_to_metadata_filter():
-    service = MagicMock()
-    service.count_memories.return_value = 1
-    service.list_memories.return_value = [
-        {
-            "id": 1,
-            "memory": "remember scope",
-            "user_id": "u01",
-            "agent_id": "a01",
-            "metadata": {"scope": "personal"},
-        }
-    ]
-
-    request = Request(
-        {
-            "type": "http",
-            "method": "GET",
-            "path": "/api/v1/memories",
-            "headers": [],
-            "query_string": b"",
-            "client": ("testclient", 50000),
-            "server": ("testserver", 80),
-            "scheme": "http",
-        }
-    )
-
-    response = await list_memories(
-        request=request,
-        user_id="u01",
-        agent_id="a01",
-        scope="personal",
-        limit=100,
-        offset=0,
-        sort_by=None,
-        order="desc",
-        time_range=None,
-        api_key="test-key",
-        service=service,
-    )
-
-    assert response.success is True
-    assert response.data["total"] == 1
-    service.count_memories.assert_called_once_with(
-        user_id="u01",
-        agent_id="a01",
-        filters={"scope": "personal"},
-    )
-    service.list_memories.assert_called_once_with(
-        user_id="u01",
-        agent_id="a01",
-        limit=100,
-        offset=0,
-        sort_by=None,
-        order="desc",
-        filters={"scope": "personal"},
     )
